UI Cleanup and refactor
