+++ styles/_typography.scss
@@ -9,7 +9,7 @@ textarea {
   color: #ccc;
   font-family: $font-family-primary, $font-family-fallback, $font-family-system;
   font-size: 16px;
-  font-weight: 300;
+  font-weight: 400;
   line-height: 1.45em;
 }
